+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 `1.9.1 (unreleased) <https://github.com/kdeldycke/meta-package-manager/compare/v1.9.0...develop>`_
 --------------------------------------------------------------------------------------------------
 
-* No changes yet.
+* Add optionnal ``version`` property on package manager definitions.
 
 
 `1.9.0 (2016-09-23) <https://github.com/kdeldycke/meta-package-manager/compare/v1.8.0...v1.9.0>`_